Fix build warnings: type conversion casts and Matrix::IsIdentity bug
- Make M_PI/DEG_TO_RAD/RAD_TO_DEG macros produce float values
- Add explicit casts for ptrdiff_t→int, size_t→uint32, LONG→float, double→float
- Fix double literal 0.03 → 0.03f in RenderCollector
- Fix Matrix::IsIdentity return statement misplaced inside inner loop

+++ W3_Jungle_Team6/Editor/Core/EditorConsole.cpp
@@ -171,7 +171,7 @@ int FEditorConsole::TextEditCallback(ImGuiInputTextCallbackData* data) {
         }
 
         if (candidates.Size == 1) {
-            data->DeleteChars(word_start - data->Buf, word_end - word_start);
+            data->DeleteChars(static_cast<int>(word_start - data->Buf), static_cast<int>(word_end - word_start));
             data->InsertChars(data->CursorPos, candidates[0]);
             data->InsertChars(data->CursorPos, " ");
         }
